Move I18n define into JavaScript file, allow webpacker supports erb
