--- app/views/layouts/_javascripts.html.erb
@@ -1,4 +1,0 @@
-<script type="text/javascript">
-  I18n.locale = "<%= I18n.locale %>"
-  I18n.translations = <%= I18n.backend.translations.to_json.html_safe %>
-</script>
