+++ app/views/layouts/application.html.erb
@@ -16,7 +16,6 @@
     <%= javascript_pack_tag 'application', media: 'all', 'data-turbolinks-track': 'reload' %>
   </head>
   <body>
-    <%= render 'layouts/javascripts' %>
     <%= render 'layouts/header' %>
     <%= render 'layouts/flash' %>
     <%= render 'layouts/bookmark_new' %>
