+++ config/webpack/environment.js
@@ -1,3 +1,6 @@
 const { environment } = require('@rails/webpacker')
+const erb = require('./loaders/erb');
+
+environment.loaders.prepend('erb', erb)
 
 module.exports = environment
