+++ config/webpacker.yml
@@ -36,6 +36,7 @@ default: &default
   extensions:
     - .mjs
     - .js
+    - .js.erb
     - .sass
     - .scss
     - .css
